# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finincial accounting API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/spendings": {
            "get": {
                "tags": [
                    "Spendings"
                ],
                "summary": "Get list of spendings",
                "description": "Get spendings",
                "operationId": "spendingsList",
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Spendings"
                ],
                "summary": "Create Spending",
                "description": "Create Spending",
                "operationId": "spendingCreate",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "date",
                                    "source_id",
                                    "type_id"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-28"
                                    },
                                    "category_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "type_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "tag_ids": {
                                        "type": "array",
                                        "items": {},
                                        "example": "[\"McDonalds\", \"BurgerKing\", \"KFC\"]"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Example description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/spendings/{id}": {
            "get": {
                "tags": [
                    "Spendings"
                ],
                "summary": "Get spending by id",
                "description": "Get spending by id",
                "operationId": "ea12106d237d98fc92faf3433dc6de85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Spending id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Spendings"
                ],
                "summary": "Update Spending",
                "description": "Update Spending by ID",
                "operationId": "spendingUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Spending ID",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-28"
                                    },
                                    "category_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "type_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "tag_ids": {
                                        "type": "array",
                                        "items": {},
                                        "example": "[\"McDonalds\", \"BurgerKing\", \"KFC\"]"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Example description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get tag by id",
                "description": "Get tag by id",
                "operationId": "d3d2e4eac468da54d8ae14faee889bb5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Delete Tag",
                "description": "Delete tag by ID",
                "operationId": "tagDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag ID",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tags"
                ],
                "summary": "Update Tag",
                "description": "Update Tag by ID",
                "operationId": "tagUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tag ID",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "McDonalds"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}